Always point to the anchor point
Previously, the code indexed into the ith location, but the
sorted array puts the point at i into the 0 position.  This
0-index point is the "origin" against which all other points
are compared, and it is invariant.

Accordingly, rather than doing a comparison from i to j, since
i changes after the array is sorted according to i, do a
comparison from 0 to j.