Sort by input points, not local points
Each time the points array is sorted, their order changes.
This truism means that it's possible that a specific point
in the array is never under the pointer i even throughout
the entire iteration.

Accordingly, to prevent this from happening, sort the pts array
by using the points in the order originally passed to
FastCollinearPoints by points